Adjust grammar and add railroad diagrams
The `grammar` folder now contains:
- the actual ABNF grammar (`chunks.abnf`),
- tools to generate a chunks document parser out of it, using APG (name of the
APG parser module: `apg-chunks.js`)
- tools to create an HTML fragment (`rr.html`) that contains all the railroad
diagrams for inclusion in the spec.

The Chunks specification itself directly includes the grammar and the railroad
diagrams files, so no need to copy them over to the specification.

To re-generate the files when `chunks.abnf` changes, run:

```js
npm run update-grammar
```

... and commit files that will have changed. The generation creates a couple of
additional files that should not be committed in the repository (they are listed
in `.gitignore` and thus will be automatically ignored)

The above command will only work provided that Node.js and Java are installed,
and provided that a `npm install` command has been issued to install needed
dependencies.