Introduce the notion of context (#29)
This update introduces the notion of a chunk context with wording and example
taken from:
https://github.com/w3c/cogai/blob/master/chunks-and-rules.md#statements-about-statements

It also describes the `@context` property and updates the "matching chunk"
algorithm accordingly (algorithm was re-worded to better account for other
corner cases. It still does not account for variables and for the `~` operator).

Linked to #25.