Handles lists, negated and wild card operators during match (#37)
* Account for lists in chunk matching algorithm

The update also clarifies that the order of atomic values in a list is
significant.

Note that the `@kindof` operator will only match if its value is an atomic
value. Rule may change in the future if that seems too strict.

Fixes #30 and #36.

* Update matching rules to account for variables, add negation operator

* Add wild card operator, update grammar and move sections

This defines the wild card operator "*", moves all name operators to a common
section and shuffle content around accordingly.

This update also refreshes the grammar to integrate name operators.