Describe mapping to RDF
Note the section explains how to map names to IRIs but does not detail full
serialization/deserialization algorithms.

The update also moves the definition of `@-properties` for conditions and
actions back to the properties section. It seems a better fit now that
there is a section that defines `@rdfmap`, `@prefix` and `@base`.